chore: Rename logging to shared utils.
