+++ data_resource/admin/routes/generator.py
@@ -1,7 +1,7 @@
 from flask_restful import Api, Resource
 from flask import Blueprint, current_app, request
 from data_resource.generator.app import start_data_resource_generator
-from data_resource.logging import LogFactory
+from data_resource.shared_utils import LogFactory
 from data_resource.admin.routes.auth_util import check_auth
 
 generator_bp = Blueprint("generator_bp", __name__)
