+++ data_resource/admin/routes/swagger.py
@@ -3,7 +3,7 @@ from flask import Blueprint
 from data_resource.db import db_session
 import data_resource.admin.models as orm
 from convert_descriptor_to_swagger import convert_descriptor_to_swagger
-from data_resource.logging import LogFactory
+from data_resource.shared_utils import LogFactory
 from data_resource.admin.routes.auth_util import check_auth
