+++ data_resource/generator/api_manager/v1_0_0/crud_functions/resource_delete.py
@@ -1,4 +1,4 @@
-from data_resource.logging import LogFactory
+from data_resource.shared_utils import LogFactory
 
 
 logger = LogFactory.get_console_logger("generator:resource-delete")
