+++ data_resource/generator/app.py
@@ -1,6 +1,6 @@
 from data_resource.generator.api_manager import generate_api
 from data_resource.generator.model_manager import create_models
-from data_resource.logging import LogFactory
+from data_resource.shared_utils import LogFactory
 import json
 from flask import current_app
 import os
