--- data_resource/logging/__init__.py
@@ -1,1 +1,0 @@
-from data_resource.logging.log_factory import LogFactory
