+++ data_resource/shared_utils/__init__.py
@@ -1,0 +1,1 @@
+from data_resource.shared_utils.log_factory import LogFactory
