+++ tests/api_manager/rest_functions/singular_resources/test_single_get.py
@@ -1,6 +1,6 @@
 from data_resource.generator.api_manager.v1_0_0.resource_handler import ResourceHandler
 from data_resource.db.base import db_session
-from data_resource.logging.api_exceptions import ApiError
+from data_resource.shared_utils.api_exceptions import ApiError
 import pytest
