+++ tests/api_manager/rest_functions/singular_resources/test_single_post.py
@@ -1,6 +1,6 @@
 from data_resource.generator.api_manager.v1_0_0.resource_handler import ResourceHandler
 import pytest
-from data_resource.logging.api_exceptions import ApiError
+from data_resource.shared_utils.api_exceptions import ApiError
 
 
 @pytest.mark.requiresdb
